fix int overlow

+++ src/main/java/io/mycat/buffer/BufferPool.java
@@ -12,8 +12,8 @@ import java.nio.ByteBuffer;
 public interface BufferPool {
     public ByteBuffer allocate(int size);
     public void recycle(ByteBuffer theBuf);
-    public int capacity();
-    public int size();
+    public long capacity();
+    public long size();
     public int getConReadBuferChunk();
     public  int getSharedOptsCount();
     public int getChunkSize();
